feat: Separate core logic into different module
