+++ pyproject.toml
@@ -5,6 +5,7 @@ description = "Semantic document search using PGAI"
 authors = ["billy <billydevbusiness@gmail.com>"]
 license = "MIT"
 readme = "README.md"
+packages = [{ include = "document_search", from = "src" }]
 
 [tool.poetry.dependencies]
 python = "^3.12"
